Refactor typography components for greater reusability
The commit contains changes in multiple components. As there were some redundancies and potential for optimized reusability in typography components, they have been adjusted accordingly. PropsWithColorAndChildren interface is renamed to ReusableItemProps and added a style prop to allow users to override the default styles. This style prop, along with the color prop, are included in ButtonText, Heading, Label, and Caption components, thus adding flexibility for the users of these components in terms of styling. This will facilitate design consistency across the app and make future modifications easier.