Add react-native-gifted-chat package to the project
The react-native-gifted-chat package version 2.4.0 was added to the package.json file. This package provides a customizable UI for implementing chat functionality in the React Native application. It was added to enhance the chat feature of the application by providing various out-of-the-box functionalities such as built-in send/receive messages, image message, bubble chat UI, etc. The change is reflected in the yarn.lock file as the new package and its dependencies are added.