Add ApolloClient configuration in utils
A new utility function has been added to configure ApolloClient. This function uses the ApolloClient constructor and sets the API URL from an environment variable. It also sets up caching with InMemoryCache. If a token is provided, it is added to the headers for authorization setup. This addition improves the way we handle GraphQL requests across the app.