Update Heading component to extend TextProps
HeadingProps interface is updated to extend TextProps. This change is made to ensure that all the properties that are applicable to the Text component in React Native can be directly used in the Heading component. Also, fontSize in h4 style in Heading/styles.ts is updated from 18 to 20 to improve readability.