Remove SendMsgFooter component from ChatScreen
This commit removes the SendMsgFooter.tsx file that was used in the ChatScreen. It became redundant after the recent restructuring of the ChatScreen's UI and therefore no longer needed. Instead, the new MessageInput component will be utilised, which provides a comprehensive and polished user experience. This removal is a part of broader codebase clean up and UI improvement initiative.